Added missing shiftIn function

+++ stm32/cores/maple/wirish_shift.cpp
@@ -25,12 +25,12 @@
 #include "wirish.h"
 
 // Source: https://github.com/arduino/Arduino/blob/master/hardware/arduino/sam/cores/arduino/wiring_shift.c
-uint32_t shiftIn( uint32_t dataPin, uint32_t clockPin, uint32_t bitOrder )
+uint8_t shiftIn(uint8_t dataPin, uint8_t clockPin, uint8_t bitOrder)
 {
 	uint8_t value = 0 ;
 	uint8_t i ;
 
-	for (i=0 ; i < 8 ; ++i) {
+	for ( i=0 ; i < 8 ; ++i ) {
 		digitalWrite( clockPin, HIGH );
 
 		if ( bitOrder == LSBFIRST ) {
